+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,9 +1,11 @@
 package ru.yandex.practicum.filmorate.controller;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+
+import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
+
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/users")
@@ -11,5 +13,17 @@ public class UserController extends ObjectController<User> {
 
     public UserController(UserService service) {
         super(service);
+    }
+
+    @PostMapping
+    @Override
+    public User create(@Valid @RequestBody User user) {
+        return service.create(user);
+    }
+
+    @PutMapping
+    @Override
+    public User put(@Valid @RequestBody User user) {
+        return service.put(user);
     }
 }
