+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
@@ -9,23 +10,16 @@ import javax.validation.constraints.Size;
 import java.time.LocalDate;
 
 @Data
+@AllArgsConstructor
 public class Film {
+
     private Integer id;
-    @NotNull(message = "Name cannot be null")
     @NotBlank(message = "Name cannot be blank")
     private String name;
     @Size(min = 1, max = 200, message = "Description must be between 1 and 200 characters")
     private String description;
     private LocalDate releaseDate;
+    @NotNull(message = "Duration cannot be null")
     @Positive(message = "Duration should be positive")
     private Integer duration;
-
-    public Film(Integer id, String name, String description,
-                LocalDate releaseDate, Integer duration) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.releaseDate = releaseDate;
-        this.duration = duration;
-    }
 }
