+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import javax.validation.constraints.Email;
@@ -9,27 +10,16 @@ import javax.validation.constraints.Past;
 import java.time.LocalDate;
 
 @Data
+@AllArgsConstructor
 public class User {
 
     private Integer id;
     @NotNull(message = "Email cannot be null")
-    @NotBlank(message = "Email cannot be blank")
     @Email(message = "Email should be valid")
     private String email;
-    @NotNull(message = "Login cannot be null")
     @NotBlank(message = "Login cannot be blank")
     private String login;
     private String name;
-    @NotNull(message = "Birthday cannot be null")
     @Past(message = "Birthday should be in the past")
     private LocalDate birthday;
-
-
-    public User(Integer id, String email, String login, String name, LocalDate birthday) {
-        this.id = id;
-        this.email = email;
-        this.login = login;
-        this.name = name;
-        this.birthday = birthday;
-    }
 }
