+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.Film;
 
 @Repository
-public class FilmRepository extends ObjectsRepository<Film> {
+public class FilmRepository extends ObjectsRepository<Integer, Film> {
 }
