+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.User;
 
 @Repository
-public class UserRepository extends ObjectsRepository<User> {
+public class UserRepository extends ObjectsRepository<Integer, User> {
 }
